+++ Flow/Changes.md
@@ -11,6 +11,7 @@
 ### New Features
 - ValueUnwraps -extensions added to utopia.flow.generic -package. By importing these extensions, you 
 can automatically unwrap values to most basic types (saves you from writing value.string or value.getInt etc.)
+- Mirror class added for supporting mapping change listening
         
 ### Fixes
 - tryMap(...) in CollectionExtensions now uses buildFrom and not Factory. This should result in better result